+++ CHANGELOG.md
@@ -10,6 +10,14 @@ and this project adheres to [Semantic Versioning][].
 
 ## [0.1.0] - tbd
 
+### Added
+
+-   Pushed `get_extent` functionality upstream to `spatialdata` (#162)
+
+### Fixed
+
+-
+
 ## [0.0.5] - 2023-10-02
 
 ### Added